requirements.txt
tensorflow-cpu==2.13.0
fastapi
uvicorn
gunicorn
python-multipart
pillow
matplotlib
numpy